feat: add Option component
