+++ src/components/Option/icon.js
