+++ src/components/Option/styled/checkmarkIcon.js
