+++ src/components/Option/styled/header.js
