+++ src/components/Option/styled/icon.js
