+++ src/components/PicklistOption/index.js
@@ -1,197 +1,14 @@
-/* eslint-disable no-script-url, react/prop-types, jsx-a11y/anchor-is-valid */
-import React, { Component } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
-import { Consumer } from '../InternalDropdown/context';
-import Icon from './icon';
-import StyledHeader from './styled/header';
-import StyledHeaderLabel from './styled/headerLabel';
-import StyledItem from './styled/item';
-import StyledIconContainer from './styled/iconContainer';
-import StyledCheckmarkIcon from './styled/checkmarkIcon';
-
-function preventDefault(event) {
-    event.preventDefault();
-}
-
-class Option extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            showOption: true,
-        };
-        this.itemRef = React.createRef();
-        this.handleClick = this.handleClick.bind(this);
-        this.handleHover = this.handleHover.bind(this);
-    }
-
-    componentDidMount() {
-        const { disabled, variant, name } = this.props;
-        const isHeader = variant === 'header';
-        if (disabled || isHeader || typeof name !== 'string') {
-            return null;
-        }
-        return this.register();
-    }
-
-    componentDidUpdate(prevProps) {
-        const { activeChildrenMap: prevActiveChildrenMap } = prevProps;
-        const { activeChildrenMap, name, variant } = this.props;
-        if (prevActiveChildrenMap !== activeChildrenMap && variant === 'default') {
-            if (activeChildrenMap[name]) {
-                this.register();
-            } else {
-                this.unregister();
-            }
-        }
-    }
-
-    componentWillUnmount() {
-        return this.unregister();
-    }
-
-    handleClick(event) {
-        const { disabled, privateOnClick, label, name, icon, value } = this.props;
-        if (disabled) {
-            event.preventDefault();
-            return null;
-        }
-        return privateOnClick(event, {
-            label,
-            name,
-            icon,
-            value,
-        });
-    }
-
-    handleHover(event) {
-        const { privateOnHover, disabled, name } = this.props;
-        if (disabled) {
-            return null;
-        }
-        return privateOnHover(event, name);
-    }
-
-    register() {
-        const { privateRegisterChild, label, name, icon, value } = this.props;
-        return setTimeout(
-            () =>
-                privateRegisterChild(this.itemRef.current, {
-                    label,
-                    name,
-                    icon,
-                    value,
-                }),
-            0,
-        );
-    }
-
-    unregister() {
-        const { privateUnregisterChild, name } = this.props;
-        if (privateUnregisterChild) {
-            return privateUnregisterChild(this.itemRef.current, name);
-        }
-        return null;
-    }
-
-    renderRightIcon() {
-        const { name, currentValueName, icon, iconPosition } = this.props;
-        const hasRightIcon = !!(icon && iconPosition === 'right');
-        if (name === currentValueName) {
-            return <StyledCheckmarkIcon />;
-        }
-        return (
-            <Icon
-                data-id="menu-item-right-icon"
-                icon={icon}
-                isVisible={hasRightIcon}
-                position={iconPosition}
-            />
-        );
-    }
-
-    render() {
-        const {
-            style,
-            className,
-            label,
-            title,
-            variant,
-            icon,
-            iconPosition,
-            disabled,
-            activeOptionName,
-            name,
-            currentValueName,
-            activeChildrenMap,
-        } = this.props;
-        const isSelected = currentValueName === name;
-        const isActive = activeOptionName === name;
-
-        if (variant === 'header') {
-            return (
-                <StyledHeader
-                    className={className}
-                    style={style}
-                    title={title}
-                    role="presentation"
-                    onMouseDown={preventDefault}
-                >
-                    <StyledHeaderLabel>{label}</StyledHeaderLabel>
-                </StyledHeader>
-            );
-        }
-
-        const showOption = !activeChildrenMap || activeChildrenMap[name];
-
-        if (showOption) {
-            const hasLeftIcon = !!(icon && iconPosition === 'left');
-
-            return (
-                <li
-                    data-selected={isSelected}
-                    className={className}
-                    style={style}
-                    role="presentation"
-                    onMouseDown={this.handleClick}
-                    onMouseEnter={this.handleHover}
-                >
-                    <StyledItem
-                        id={name}
-                        role="option"
-                        aria-selected={isActive}
-                        aria-disabled={disabled}
-                        ref={this.itemRef}
-                        isSelected={isSelected}
-                        isActive={isActive}
-                    >
-                        <StyledIconContainer title={title}>
-                            <Icon
-                                data-id="menu-item-left-icon"
-                                icon={icon}
-                                isVisible={hasLeftIcon}
-                                position={iconPosition}
-                            />
-
-                            {label}
-                        </StyledIconContainer>
-                        {this.renderRightIcon()}
-                    </StyledItem>
-                </li>
-            );
-        }
-        return null;
-    }
-}
+import Option from '../Option';
 
 /**
  * Represents a list options in a menu.
  * @category Form
  */
 export default function PicklistOption(props) {
-    return <Consumer>{values => <Option {...props} {...values} />}</Consumer>;
+    return <Option {...props} />;
 }
-
-export { Option };
 
 PicklistOption.propTypes = {
     /** Text of the PicklistOption. */
