+++ src/components/index.js
@@ -41,6 +41,7 @@ export { default as MenuItem } from './MenuItem';
 export { default as Modal } from './Modal';
 export { default as MonthlyCalendar } from './MonthlyCalendar';
 export { default as Notification } from './Notification';
+export { default as Option } from './Option';
 export { default as Pagination } from './Pagination';
 export { default as Picklist } from './Picklist';
 export { default as PicklistOption } from './PicklistOption';
